refactor[#150]: model 삭제 및 repo dto로 경로 수정
