+++ src/main/java/com/plog/plogbackend/domain/map/controller/dto/response/MapCountResponse.java
@@ -1,6 +1,6 @@
 package com.plog.plogbackend.domain.map.controller.dto.response;
 
-import com.plog.plogbackend.domain.map.model.MapCount;
+import com.plog.plogbackend.domain.map.repository.dto.MapCount;
 import lombok.Builder;
 
 @Builder
