+++ src/main/java/com/plog/plogbackend/domain/map/repository/dto/PlaceDetail.java
@@ -1,4 +1,4 @@
-package com.plog.plogbackend.domain.map.model;
+package com.plog.plogbackend.domain.map.repository.dto;
 
 import com.plog.plogbackend.domain.post.enums.PlaceCategoryCode;
 import lombok.AccessLevel;
