+++ src/main/java/com/plog/plogbackend/domain/map/service/MapService.java
@@ -1,13 +1,13 @@
 package com.plog.plogbackend.domain.map.service;
 
 import com.plog.plogbackend.domain.bookmark.repository.BookMarkRepository;
-import com.plog.plogbackend.domain.map.model.MapCount;
-import com.plog.plogbackend.domain.map.model.MapPin;
-import com.plog.plogbackend.domain.map.model.PlaceDetail;
-import com.plog.plogbackend.domain.map.model.PlaceRecord;
-import com.plog.plogbackend.domain.map.model.PlaceSearchResult;
-import com.plog.plogbackend.domain.map.model.PlaceSummary;
-import com.plog.plogbackend.domain.map.model.Viewport;
+import com.plog.plogbackend.domain.map.repository.dto.MapCount;
+import com.plog.plogbackend.domain.map.repository.dto.MapPin;
+import com.plog.plogbackend.domain.map.repository.dto.PlaceDetail;
+import com.plog.plogbackend.domain.map.repository.dto.PlaceRecord;
+import com.plog.plogbackend.domain.map.repository.dto.PlaceSearchResult;
+import com.plog.plogbackend.domain.map.repository.dto.PlaceSummary;
+import com.plog.plogbackend.domain.map.repository.dto.Viewport;
 import com.plog.plogbackend.domain.map.repository.MapQueryRepository;
 import com.plog.plogbackend.domain.member.repository.MemberRepository;
 import com.plog.plogbackend.domain.post.repository.PostRepository;
